Policy: MOVEONLY: Move CFeeRate to policy
