+++ src/miner.cpp
@@ -15,6 +15,7 @@
 #include "hash.h"
 #include "main.h" // cs_main
 #include "net.h"
+#include "policy/feerate.h"
 #include "policy/policy.h"
 #include "primitives/block.h"
 #include "primitives/transaction.h"
