+++ src/wallet/wallet.h
@@ -7,6 +7,7 @@
 #define BITCOIN_WALLET_H
 
 #include "amount.h"
+#include "policy/feerate.h"
 #include "primitives/transaction.h"
 #include "tinyformat.h"
 #include "ui_interface.h"
